--- app/src/main/java/bg/zahov/app/data/interfaces/WorkoutActionHandler.kt
@@ -1,9 +1,0 @@
-package bg.zahov.app.data.interfaces
-
-import kotlinx.coroutines.flow.StateFlow
-
-interface WorkoutActionHandler {
-    val shouldFinish: StateFlow<Boolean>
-    fun tryToFinish()
-    fun completeFinishAttempt()
-}
