+++ app/src/main/java/bg/zahov/app/data/interfaces/WorkoutProvider.kt
@@ -16,6 +16,11 @@ interface WorkoutProvider {
      * observable exercise that was clicked in history screen
      */
     val clickedPastWorkout: StateFlow<HistoryInfoWorkout>
+    /**
+     * A [StateFlow] indicating whether a finish attempt has been signaled.
+     * The value is `true` when a finish attempt is in progress, and `false` otherwise.
+     */
+    val shouldFinish: StateFlow<Boolean>
     suspend fun getTemplateWorkouts(): Flow<List<Workout>>
     suspend fun getPastWorkouts(): Flow<List<Workout>>
     suspend fun addTemplateWorkout(newWorkout: Workout)
@@ -67,4 +72,14 @@ interface WorkoutProvider {
      * Returns the last performed workout if any
      */
     fun getLastWorkout(): HistoryWorkout?
+
+    /**
+     * Attempts a finish attempt for the workout
+     */
+    fun tryToFinish()
+
+    /**
+     * Completes the finish attempt for the workout
+     */
+    fun completeFinishAttempt()
 }
