+++ app/src/main/java/bg/zahov/app/data/provider/WorkoutProviderImpl.kt
@@ -46,6 +46,9 @@ class WorkoutProviderImpl : WorkoutProvider {
     override var clickedPastWorkout: StateFlow<HistoryInfoWorkout> =
         MutableStateFlow(HistoryInfoWorkout())
 
+    private val _shouldFinish = MutableStateFlow(false)
+    override val shouldFinish: StateFlow<Boolean>
+        get() = _shouldFinish
 
     private val workoutRepo = WorkoutRepositoryImpl.getInstance()
     private val errorHandler = ServiceErrorHandlerImpl.getInstance()
@@ -56,6 +59,20 @@ class WorkoutProviderImpl : WorkoutProvider {
      */
     override fun getLastWorkout(): HistoryWorkout? {
         return lastWorkoutPerformed?.toHistoryWorkout()
+    }
+
+    /**
+     * Signals the start of a finish operation by setting `_shouldFinish` to `true`.
+     */
+    override fun tryToFinish() {
+        _shouldFinish.value = true
+    }
+
+    /**
+     * Completes the finish operation by resetting `_shouldFinish` to `false`.
+     */
+    override fun completeFinishAttempt() {
+        _shouldFinish.value = false
     }
 
     override suspend fun getTemplateWorkouts(): Flow<List<Workout>> =
@@ -148,7 +165,8 @@ class WorkoutProviderImpl : WorkoutProvider {
     override suspend fun getExerciseHistory(): Flow<List<ExerciseHistoryInfo>> = exerciseHistory
     override suspend fun <T> getPreviousWorkoutState(): T? = null
 
-    override suspend fun <T> addWorkoutState(realmWorkoutState: T) { /* TODO() */ }
+    override suspend fun <T> addWorkoutState(realmWorkoutState: T) { /* TODO() */
+    }
 
     override suspend fun updateTemplateWorkout(
         workoutId: String,
@@ -230,7 +248,11 @@ fun Workout.toStartWorkout(): StartWorkout = StartWorkout(
  */
 fun Exercise.toStartWorkoutExercise(): StartWorkoutExercise = StartWorkoutExercise(
     name = this.name,
-    exercise = if (this.sets.isNotEmpty()) { "${this.sets.size} x " } else { "" } + this.name,
+    exercise = if (this.sets.isNotEmpty()) {
+        "${this.sets.size} x "
+    } else {
+        ""
+    } + this.name,
     bodyPart = this.bodyPart,
     category = this.category
 )
