+++ app/src/main/java/bg/zahov/app/ui/workout/WorkoutScreen.kt
@@ -158,7 +158,6 @@ fun WorkoutScreen(
 
     if (state.isFinished) {
         LaunchedEffect(Unit) {
-            workoutViewModel.resetFinishTrigger()
             onFinish()
         }
     }
